+++ app/src/main/java/org/sopt/linkareer/feature/certification/CertificationGuideRoute.kt
@@ -1,6 +1,5 @@
-package org.sopt.linkareer.feature.certification.navigation
+package org.sopt.linkareer.feature.certification
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -8,28 +7,25 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.linkareer.R
-import org.sopt.linkareer.core.designsystem.theme.Gray700
 import org.sopt.linkareer.core.designsystem.theme.Gray900
-import org.sopt.linkareer.core.designsystem.theme.LINKareerAndroidTheme
 import org.sopt.linkareer.core.designsystem.theme.White
 import org.sopt.linkareer.core.designsystem.theme.defaultLINKareerTypography
 import org.sopt.linkareer.feature.certification.component.CertificationConfirmButton
+import org.sopt.linkareer.feature.certification.component.CertificationGuideList
 import org.sopt.linkareer.feature.certification.component.CertificationTopBar
 
 @Composable
-fun CertificationCheckSuccessRoute() {
+fun CertificationGuideRoute() {
+    CertificationGuideScreen()
 }
 
 @Composable
-fun CertificationCheckSuccessScreen() {
+fun CertificationGuideScreen() {
     Column(
         modifier =
             Modifier
@@ -43,32 +39,22 @@ fun CertificationCheckSuccessScreen() {
                     .padding(top = 27.dp),
         )
         Text(
-            text = stringResource(R.string.certification_check_title),
+            text = stringResource(R.string.certification_guide_title),
             style = defaultLINKareerTypography.title3B16,
             color = Gray900,
             modifier =
                 Modifier
                     .padding(start = 17.dp, top = 16.dp),
         )
-        Text(
-            text = stringResource(R.string.certification_check_subtitle),
-            style = defaultLINKareerTypography.body10M11,
-            color = Gray700,
+        CertificationGuideList(
             modifier =
                 Modifier
-                    .padding(start = 17.dp, top = 8.dp),
-        )
-        Image(
-            imageVector = ImageVector.vectorResource(R.drawable.ic_check_circle_144),
-            contentDescription = stringResource(R.string.certification_check_success),
-            modifier =
-                Modifier
-                    .align(Alignment.CenterHorizontally)
-                    .padding(top = 147.dp),
+                    .padding(horizontal = 17.dp)
+                    .padding(top = 32.dp),
         )
         Spacer(modifier = Modifier.weight(1f))
         CertificationConfirmButton(
-            buttonText = R.string.certification_confirm_button,
+            buttonText = R.string.certification_ok_button,
             onClickButton = {},
             modifier =
                 Modifier
@@ -80,8 +66,6 @@ fun CertificationCheckSuccessScreen() {
 
 @Preview(showBackground = true)
 @Composable
-fun CertificationCheckSuccessScreenPreview() {
-    LINKareerAndroidTheme {
-        CertificationCheckSuccessScreen()
-    }
+fun CertificationGuideScreenPreview() {
+    CertificationGuideScreen()
 }
